+++ NEWS.md
@@ -9,6 +9,8 @@
   * Moved all `require()` in individual files to `testthat.R`. This removed many `require()` that were not needed.
 * Fixed four minor errors in documentation from legacy uses of `\R{}` rather than `\code{}`.
 
+* `depletion()`: Modified.
+  * Added formula notation such that `depletion()` wash changed to a method call and `depletion.default()` and `depletion.formula()` were added. Tests for the formula were included.
 * `GompertzFuns()`: Accepted pull request related to [#112](https://github.com/fishR-Core-Team/FSA/issues/112) that fixed several typos and dead links in the documentation ... thanks Arni. Corrected the erroneous reference to t* (should have been t0) in the documentation for the Gompertz function (fixes [#113](https://github.com/fishR-Core-Team/FSA/issues/113) ... thanks again to Arni).
 * `metaM()`: Modified to address [#114](https://github.com/fishR-Core-Team/FSA/issues/114).
   * Returns data.frame rather than list.